refactor server effective policy simulation service
